+++ telekit/_init.py
@@ -25,6 +25,7 @@ from ._inline_buttons import InlineButton
 from ._user import User
 from .senders import BaseSender
 from ._state import TelekitState
+from .chat import Chat
 
 import telebot
 
@@ -33,6 +34,7 @@ __all__ = ["init"]
 def init(bot: telebot.TeleBot) -> None:
     TelekitState._init(bot)
     BaseSender._init(bot)
+    Chat._init(bot)
     Handler._init(bot)
     Chain._init(bot)
     InputHandler._init(bot)