+++ telekit/debug.py
@@ -2,6 +2,7 @@ class Debug:
     timeout_warnings: bool = False
     deletion_warnings: bool = False
     callback_query_tracing: bool = False
+    duplicate_handler_warnings: bool = True
 
     @classmethod
     def set_all(cls, value: bool) -> None: